separate hooks from logic
