+++ src/frontend/interfaces/desktop/App.scss
@@ -2,3 +2,13 @@ body {
   background-color: #ddd;
   margin: 0;
 }
+
+h1 {
+  font-size: 2rem;
+}
+.p-inputtext {
+  background: #fafafa;
+}
+td {
+  border: 1px solid #ddd;
+}
