+++ src/frontend/interfaces/desktop/components/LandingPage/TournamentInfo/TournamentInfo.scss
@@ -1,5 +1,4 @@
 .TournamentInfo {
-  background: red;
   display: flex;
   flex-direction: column;
   align-items: center;
@@ -7,7 +6,6 @@
   width: 100%;
   height: 100%;
   padding: 10px;
-  background-color: #f5f5f5;
   border-radius: 10px;
   box-shadow: 0 0 10px 0 rgba(0, 0, 0, 0.1);
   table {
@@ -17,11 +15,10 @@
       text-align: left;
       padding: 10px;
       background-color: #f5f5f5;
-      border-bottom: 1px solid #ddd;
     }
     td {
       padding: 10px;
-      border-bottom: 1px solid #ddd;
+      border: none;
     }
   }
 }
